Added Owned by Namespace field to Trigger Handler so HEDA and other packages can isolate their Trigger Handlers for the purpose of default management. Also added the Owned by Namespace field to the TDTM_Token class.
Implemented hashCode() and equals() methods for the TDTM_Token custom type of the TDTM_Global_API class. The enables the use of sets helping to optimize the default configuration comparison algorithm. Added a handlerId field to the Token class to track when a handler could be updated instead of inserted into the database.

Created TDTM_Manager class. This class will be instantiated when updating or installing new default TDTM triggers. Upon instantiation the TDTM_Manager will contain the existing configuration and you can pass in a Set of tokens to compare against the current configuration to update/insert/delete. The TDTM_Manager class is

There is also a simple method on TDTM_Manager which accepts a map of Trigger Handlers and returns a TDTM_Token SET

The install script was modified to target upgrades to 1.36 so the Owned by Namespace field can be utilized moving forward for HEDA managed trigger handlers.